Move syncers and changesets out of event
