+++ lib/code_corps/github/event/issues.ex
@@ -10,12 +10,13 @@ defmodule CodeCorps.GitHub.Event.Issues do
   alias CodeCorps.{
     GitHub.Event.Common.RepoFinder,
     GitHub.Event.Issues.IssueLinker,
-    GitHub.Event.Issues.TaskSyncer,
     GitHub.Event.Issues.Validator,
+    GitHub.Sync.Issue.Task,
     Repo,
     Task
   }
-  alias CodeCorps.GitHub.Syncers.User.RecordLinker, as: UserRecordLinker
+  alias CodeCorps.GitHub.Sync.Issue.Task, as: IssueTaskSyncer
+  alias CodeCorps.GitHub.Sync.User.RecordLinker, as: UserRecordLinker
   alias Ecto.Multi
 
   @type outcome :: {:ok, list(Task.t)} |
@@ -53,7 +54,7 @@ defmodule CodeCorps.GitHub.Event.Issues do
     |> Multi.run(:repo, fn _ -> RepoFinder.find_repo(payload) end)
     |> Multi.run(:issue, fn %{repo: github_repo} -> link_issue(github_repo, payload) end)
     |> Multi.run(:user, fn %{issue: github_issue} -> UserRecordLinker.link_to(github_issue, payload) end)
-    |> Multi.run(:tasks, fn %{issue: github_issue, user: user} -> github_issue |> TaskSyncer.sync_all(user, payload) end)
+    |> Multi.run(:tasks, fn %{issue: github_issue, user: user} -> github_issue |> IssueTaskSyncer.sync_all(user, payload) end)
     |> Repo.transaction
     |> marshall_result()
   end
